even more refactoring
