+++ scripts/gan_extension.py
@@ -1,4 +1,9 @@
-import os
+# from __future__ import annotations
+# from typing import Union
+# from pathlib import Path
+# import numpy as np
+# from PIL import Image
+
 from modules import scripts #, shared
 import lib_gan_extension.ui
